backend/ml_service/requirements.txt
# Python ML Service Requirements
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
matplotlib==3.8.2
seaborn==0.13.0
tensorflow==2.15.0
requests==2.31.0